// FrontEnd/src/interfaces/project.interface.ts
export interface ProjectsI {
  _id: string
  title: string
  slug: string
  description: string
  category: string
  technologies: string[]
  features: string[]
  githubUrl: string
  liveUrl: string
  thumbnail: string
  liveDemo: string
  featured: boolean
  order: number
}


export interface SpeceficProjectI {
  _id: string
  title: string
  slug: string
  description: string
  category: string
  technologies: string[]
  features: string[]
  githubUrl: string
  liveUrl: string
  thumbnail: string
  liveDemo: string
  featured: boolean
  order: number
}